refactored FeedDownloader to take JSON file as input
FeedDownloader now takes a JSON file path as a list of all RSS feeds to parse. It downloads each rss and iterates through the items. Downloads the item if the link is not already inserted. app.js now takes a parameter to run method either as cronjob or nor. Feed Schema is updated to contain one unique link, with an array of group strings to indicate how many rss feeds contain this news.