continue to next item if feed is already inserted
I used to skip to the next rss list if the current item is already added, assuming that the following items in the list would have been already added. However, debugging showed that I generally stop the process in the middle of running, so I miss items in the list. Hence the change...